Add files via upload
Monster Finder is a captivating web application built with JavaScript that offers users the ability to search and discover monsters by their names. Whether you're a fan of mythical creatures or an avid gamer seeking to gather information on different monsters, Monster Finder provides an intuitive and user-friendly platform for exploring the vast world of these fascinating beings.